Use code splitting to make the library load faster (#2644)
