+++ src/ui/components/Dashboard/Dashboard.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { connect, ConnectedProps } from "react-redux";
 const DashboardViewer = require("./DashboardViewer").default;
 const Loader = require("../shared/Loader").default;
-import { selectors } from "ui/reducers";
+import * as selectors from "ui/reducers/app";
 
 import { UIState } from "ui/state";
 import hooks from "ui/hooks";
