+++ src/ui/components/Dashboard/Navigation/WorkspaceDropdownButton.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { connect, ConnectedProps } from "react-redux";
 import useAuth0 from "ui/utils/useAuth0";
-import { selectors } from "ui/reducers";
+import * as selectors from "ui/reducers/app";
 import { UIState } from "ui/state";
 import { Workspace } from "ui/types";
 import { ChevronDownIcon } from "@heroicons/react/solid";
