+++ src/ui/utils/telemetry.ts
@@ -1,5 +1,38 @@
 import mixpanel from "mixpanel-browser";
 import * as Sentry from "@sentry/react";
+import { Integrations } from "@sentry/tracing";
+import { skipTelemetry } from "./environment";
+
+export function setupTelemetry(context: Record<string, any>) {
+  const ignoreList = ["Current thread has paused or resumed", "Current thread has changed"];
+  mixpanel.init("ffaeda9ef8fb976a520ca3a65bba5014");
+
+  if (skipTelemetry()) {
+    mixpanel.disable();
+    return;
+  }
+
+  Sentry.init({
+    dsn: "https://41c20dff316f42fea692ef4f0d055261@o437061.ingest.sentry.io/5399075",
+    integrations: [new Integrations.BrowserTracing()],
+    tracesSampleRate: 1.0,
+    release: process.env.REPLAY_RELEASE ? process.env.REPLAY_RELEASE : "development",
+    beforeSend(event) {
+      if (event) {
+        const exceptionValue = event?.exception?.values?.[0].value;
+        if (ignoreList.some(ignore => exceptionValue?.includes(ignore))) {
+          return null;
+        }
+      }
+
+      return event;
+    },
+  });
+
+  mixpanel.register({ recordingId: context.recordingId });
+
+  Sentry.setContext("recording", { ...context, url: window.location.href });
+}
 
 export function setTelemetryContext(
   userId: string | undefined,
